move fast logic to controller.py and handler.py
