+++ examples/doornveld.py
@@ -6,7 +6,6 @@ import logging
 import sys
 
 from velbusaio.controller import Velbus
-from velbusaio.velbusfast import VelbusFast
 
 parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
 parser.add_argument("--connect", help="Connection string", required=True)
@@ -15,8 +14,7 @@ args = parser.parse_args()
 
 
 async def main(connect_str: str):
-    # velbus = Velbus(dsn=connect_str)
-    velbus = VelbusFast(destination=connect_str)
+    velbus = Velbus(dsn=connect_str)
     cache_dir = velbus.get_cache_dir()
     print(f"Cache dir: {cache_dir}")
